Rework documentation loading mechanism
The documentation registry is now a fully fledged Forge registry, which
means that DYMM can leverage features such as automatic checking of
registry names, overrides, prefix logging (damn it) and so on.

Also, now each ModDocumentation object can store references to multiple
ItemStacks, which eases not only the amount of objects created, but also
cross-doc references in preparation for requirements, which will be
added in the near future.

A stub for factories has also been added, which allows them to be
defined through JSON and not only through code.

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>