Start migration towards newer framework for DYM
First of all, there's a new Logger that can be used and that supports
some sort of markers. Also, this will allow us to not depend on a single
logger instance but have multiple, one per class.

Also, our Mod ID is currently 'dym' so it makes no sense to keep the old
package. Also, I am not currently aware of any mods that depend on
DYMM's API, so this transition shouldn't cause any issues.

Last but not least, this also allows us to have a perfectly clean code
base to start building upon, without having any issues with leftovers.
Code will be ported slowly from one side to the other and bridges will
be applied (like now to let the mod actually load).

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>