Shade Loader and Kotlin Bridge libraries
Loader is a component of Boson that was written in Kotlin for Kotlin,
and as such it leverages certain features that are present only in
Kotlin, such as nullable checks and when-expressions. Also, the main
goal of the Boson Loader was to provide an efficient and goal-agnostic
API to load files: this is why we decided to use this as a library.
Also, DYMM wants to remain as lightweight as possible in terms of
dependencies, so we would have had to relocate the library anyway.
Leveraging this, the entire code was rewritten from Kotlin to Java.

The Kotlin Bridge library is a Work-in-progress library (being developed
in a closed-source manner) that aims to bring some Kotlin features to
Java. In this case, its usage is very important because it allows us to
use mostly unaltered Kotlin code (at least in its main form) in Java.

Anyway, the main reason for this library is to allow us to swap
processing, locators, reporters etc in a very indipendent fashion,
without having to constantly modify and edit the source code of the
Loader itself, but just the processing, locating, reporting bits.

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>