Add basic module for JSON loading
While this doesn't actually let anybody load JSON because target and
condition factories are completely missing and no consumers are
currently registered and provided by this mod, it should still work.

Of course, as the JSON spec implies, compatibility with the previous
JSON files is broken, but they will get converted anyway as part of the
update process.

Signed-off-by: TheSilkMiner <thesilkminer@outlook.com>